feat: add support for creating GitHub Actions secrets
This commit adds support for creating GitHub Actions secrets using the
`github_actions_secret` resource. The secrets are defined in the
`repositories` variable as a map of maps, where each repository can have
a map of secrets. The secrets are flattened into a list of objects and
then transformed into a map for use in the `github_actions_secret`
resource.

The `github_actions_secret` resource is created for each secret in each
repository. The secret name is converted to lowercase to comply with
GitHub's requirements.